--HG--
branch : map-report


Former-commit-id: 06d15ed53c670c484c96be8c124996113d5a1154

+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Views/Questionnaire/QuestionnaireBrowseInputModel.cs
@@ -79,7 +79,7 @@ namespace WB.Core.SharedKernels.DataCollection.Views.Questionnaire
 
         public Guid? QuestionnaireId { get; set; }
 
-        public int? Version { get; set; }
+        public long? Version { get; set; }
 
         #endregion
     }
